+++ pycatia/cat_tps_interfaces/non_semantic_dimension.py
@@ -8,6 +8,7 @@
         and thus help debugging in pycatia.
         
 """
+import inspect
 
 from pycatia.drafting_interfaces.drawing_dimension import DrawingDimension
 from pycatia.system_interfaces.any_object import AnyObject
@@ -16,6 +17,9 @@ from pycatia.cat_tps_interfaces.dimension_limit import DimensionLimit
 
 class NonSemanticDimension(AnyObject):
     """
+
+    Introduced in V5-6R2017.
+
         .. note::
             :class: toggle
 
@@ -33,6 +37,12 @@ class NonSemanticDimension(AnyObject):
     """
 
     def __init__(self, com_object):
+        self.release_check(
+            self.application.system_configuration.release,
+            27,
+            f'{self.__class__.__name__}.{inspect.stack()[0][3]}',
+        )
+
         super().__init__(com_object)
         self.non_semantic_dimension = com_object