+++ pycatia/cat_tps_interfaces/semantic_gdt.py
@@ -30,6 +30,9 @@ from pycatia.cat_tps_interfaces.tolerance_zone import ToleranceZone
 
 class SemanticGDT(AnyObject):
     """
+
+    Introduced in V5-6R2017.
+
         .. note::
             :class: toggle
 
@@ -47,6 +50,11 @@ class SemanticGDT(AnyObject):
     """
 
     def __init__(self, com_object):
+        self.release_check(
+            self.application.system_configuration.release,
+            27,
+            f'{self.__class__.__name__}.{inspect.stack()[0][3]}',
+        )
         super().__init__(com_object)
         self.semantic_gdt = com_object