+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * updated Part.annotation_sets return type to be AnnotationSets in lieu of the
   generic Collection object.
 * updated Annotations.item2 to return an Annotation2 object. Was AnyObject. Issue #236.
+* added CATIA version checks for Object introduced in V5-6R2017 (R27).
 
 ## 0.8.0